perf(balancer) load upstreams before cleaning cache
When there are a large number of upstreams in the database or the
database is under heavy load, loading upstreams will take longer and,
sometimes, unacceptably longer. This change address said issue by
keeping the stale cache until the new data is loaded from the database.
As there might be stale content with this change, this behavior only
exists when worker consistency is set to eventual.